requirements.txt
cheroot==8.5.2
CherryPy==18.6.1
graphviz==0.19
jaraco.classes==3.2.1
jaraco.collections==3.3.0
jaraco.functools==3.3.0
jaraco.text==3.5.0
joblib==1.1.0
more-itertools==8.8.0
numpy==1.21.4
portend==2.7.1
psutil==5.8.0
pycrc==0.9.2
pycryptodomex==3.11.0
pytz==2021.1
scapy==2.4.5rc1.dev215
scikit-learn==1.0.1
scipy==1.7.2
six==1.16.0
tempora==4.1.1
threadpoolctl==3.0.0
zc.lockfile==2.0
